Add basic caching for address and weather data
I considered a number of approaches before settling on this commit.
While it is possible to add page-level caching on the `weather#show`
route (and it is by-far the most performant), it comes with
some downsides:

 - Page content _must_ be the same for all users, so showing any dynamic
   data or using any part of the Rails backend after caching is
   impossible (by design)

 - Cache expiration is awkward, since entries always end up on the
   filesystem. You must walk the fileystem to expire 30 minutes _after_
   creation or enqueue a job to run later and delete the entry

 - Sharing cache data is potentially more difficult depending on
   deployment scenarios—if there are many instances of this app
   they must be configured to share a networked volume for the cache

The requirements mention _real-world_ usage, and page-level caching is
not a typical feature of Rails apps I've worked on. Instead, we
typically rely on the Rails.cache for fine-grained caching.

So I've added caching in two spots:

 1. Geocoder is configured to cache address lookups automatically
    (indefinitely)

 2. The weather#show action is cached at the action-level, for 30
    minutes

Cache data is stored in SQLite per current configuration. This is _not_
something I'd recommend in production but works all the same for
exemplary purposes. Ideally something that can be deployed independent
of the primary database is used, like another DB, redis, or memcached.