Remove support for canada and display cache status
While testing my solution I realized there are two issues. First, I
was not guaranteed to get an accurate result from Nominatim while
searching for zip/postal codes using an unstructured query.

Nominatim provides a way to search for postal/zip code alone, but it
is not supported by the Geocoder gem. I ended up extending Geocoder
with a new lookup that adds basic support for this.

During testing I also noticed that a number of Canadian postal codes
are simply not returned in results (at least where I am in Toronto).
Turns out this is a known, rather long-standing issue with the API.

https://github.com/osm-search/Nominatim/issues/2410 has more
information.

Rather than go down the path of fixing this I removed the functionality
altogether. It is not a requirement in the specification and will
require further investigation (and possible a different API provider in
Geocoder) to support properly.

I also added a simple cache indicator including both the time of cache
generation along with a boolean indicating whether or not the result
is from the cache.